move updateSelectable to selection service
