+++ packages/ag-grid-community/src/context/context.ts
@@ -450,7 +450,6 @@ export type BeanName =
     | 'rowRenderer'
     | 'rowStyleService'
     | 'scrollVisibleService'
-    | 'selectableService'
     | 'selectionController'
     | 'selectionHandleFactory'
     | 'selectionService'
