+++ packages/ag-grid-enterprise/src/rowGrouping/groupStage/groupStage.ts
@@ -1,8 +1,14 @@
-import type { BeanCollection, GridOptions, IRowNodeStage, NamedBean, StageExecuteParams } from 'ag-grid-community';
+import type {
+    BeanCollection,
+    GridOptions,
+    IRowNodeStage,
+    ISelectionService,
+    NamedBean,
+    StageExecuteParams,
+} from 'ag-grid-community';
 import { BeanStub, ClientSideRowModelSteps } from 'ag-grid-community';
 
 import { GroupStrategy } from './groupStrategy/groupStrategy';
-import type { SelectableService } from './selectableService';
 import { TreeStrategy } from './treeStrategy/treeStrategy';
 
 export class GroupStage extends BeanStub implements NamedBean, IRowNodeStage {
@@ -17,11 +23,11 @@ export class GroupStage extends BeanStub implements NamedBean, IRowNodeStage {
     ]);
     public step: ClientSideRowModelSteps = ClientSideRowModelSteps.EVERYTHING;
 
-    private selectableService: SelectableService;
+    private selectionService: ISelectionService | undefined;
     private strategy: GroupStrategy | TreeStrategy | undefined;
 
     public wireBeans(beans: BeanCollection) {
-        this.selectableService = beans.selectableService as SelectableService;
+        this.selectionService = beans.selectionService;
     }
 
     public execute(params: StageExecuteParams): void {
@@ -36,7 +42,7 @@ export class GroupStage extends BeanStub implements NamedBean, IRowNodeStage {
 
         strategy.execute(params);
 
-        this.selectableService.updateSelectableAfterGrouping();
+        this.selectionService?.updateSelectable(true);
     }
 
     public override destroy(): void {
